+++ src/Proton.Server.Resource/Authentication/Scripts/AuthenticationScript.cs
@@ -82,7 +82,7 @@ public class AuthenticationScript : IStartup
             await account.Register(p.SocialClubName);
         }
 
-        int id = await account.Login(p.Ip);
+        long id = await account.Login(p.Ip);
         if(id != 0)
         {
             p.ProtonId = id;
